[Loaded] Make IsLoaded it's own variable.
Don't make IsLoaded mirror the IsAttached state as attaching and loading
are actually two distinct events and the order in which they happen
matters. We need to trigger some stuff when the element is loaded which
happens after ElementAdded is called. If we trigger on IsAttached then
we'll do our work before ElementAdded is called and then break.