Introduce bitmap caching.
An element's bitmap cache is invalidated whenever
UIElement::Invalidate is called so to make this
useful all changes that doesn't require the cache
to be invalidated must call UIElement::Invalidate
on the parent instead of on itself.

* grid.cpp:
* runtime.cpp:
* stylus.cpp:
* uielement.h:
* uielement.cpp: Have GetCr return NULL for contexts with
pre-rendered bitmaps and in that way prevent unnecessary
rendering of cached elements.

* uielement.h:
* uielement.cpp: Add initial bitmap cache support.
New functions InvalidateBitmapCache and InvalidateCacheMode.

* runtime.h:
* runtime.cpp: Remove ContextNode::GetTarget.
Add ContextNode::GetBitmap and ContextNode::SetBitmap.