fix a couple of issues with SL4 sample browser.
kind of gross, but sendermachine errors need to be reported
asyncronously.  The sample browser was re-sending the message in the
error handler, so we blow the stack.

Also, the browser was creating a listener in local scope but the sender
in global scope.  So when sending a message we don't find the sender
properly.  Fall back to the local/domain scope if we don't find the
listener in global scope.  This does NOT sound like correct behavior to
me, but it's required for the browser to work.