[CollectionView] Propagate changes using INotifyCollectionChanged
When either sorting or filtering is enabled, ensure that the internal
collections we use implement INotifyCollectionChanged so that it becomes
trivial to keep things in sync when adding/removing items from Groups or
the filtered list. This is needed for the case where one item is added
to multiple groups.