Revert aed492 and fix it the right way
The problem was that we were clearing the managed ref to the delegate
when we called NativeMethods.event_object_remove_handler.  This native
method doesn't *always* remove the handler, however (if you're in the
middle of emitting the event).  So we'd be left with a pointer to a
GC'ed delegate in unmanaged land.

The fix is to have the unmanaged side drive the process.  The managed
side still calls through that native method, but it (the managed side)
no longer clears the managed ref to the delegate at the same time.
Instead we register a GDestroyNotify handler to do the actual removal.
This is either done synchronously when we're not emitting, or it's done
at the end of the emit if we are.