Don't invoke data_dtors registered from managed land when the DO is destroyed
This might take a little explanation:

the previous change to events registers a data_dtor so that managed code
can be notified by unmanaged when the event handler is removed.  This is
the only way to guarantee that the managed structures we need for the
event live long enough but no longer, while still working with the
pending_removal behavior we need.

The trouble is, we were invoking data_dtors both at the time when the
EventClosure is removed (which is fine for both unmanaged and managed
delegates) as well as when the DO is destroyed.  This works fine for
C++, but the trouble for managed code is this.  For the DO to have gone
away, its entire graph of references must have already gone away (or be
slated for finalization).  If the order is such that the DO is finalized
before the delegates, we win.  But if the order is delegates first, we
lose (and abort).

So, the fix is to add a flag to AddHandler to turn on
"data_dtor_invocation_at_destroy_time"  (it defaults to true, so
unmanaged code doesn't need to change.)  Managed code sets it to false.
Also we explicitly invoke the data_dtor whenever a closure is removed
from the list (and set it to null so that we don't double invoke in
~EventClosure).

see kids?  garbage collection is fun!