[CollectionView] Make CollectionViewGroup implement INotifyCollectionChanged
Create an implementation of CollectionViewGroup which implements
INotifyCollectionChanged. Use this as the root element of the group
subtree and make it control how elements are added/removed from the
group subtree. This greatly simplifies the logic required to keep things
in sync when the CollectionView needs to emit INotifyCollectionChanged
events.