change the way we deal with inherited properties, highly optimizing property lookup
at every point in the tree where an inheritable property is valid, we
store a reference to the next element up the tree with that property set
with a higher precedent.  I.e. if there's a node up the tree with a
local value, that's the node we point to.

This way we can reduce the number of lookups in the upward tree walk
from N (where N is the number of nodes between GetValue and the source)
to <= 1.  if there's no node further up the tree, we do 0 walks.  For a
particularly contrived testcase, we see a 1000% improvement.

we still do a tree walk downward whenever we change the value of an
inheritable property, but given that lookups are so much faster, we end
up spending less time during this walk as well (in the contrived
test, it's merely a ~300% improvement).