fix propagation and clearing of inherited properties.
we need to deal with the case of elements existing between elements that
inherit properties.  we basically do this by propagating the property
everywhere, even if the element won't use it.  that way when adding a
subtree, all the properties we're interested in exist on rootParent with
proper references to the source.

also clear inherited properties on the visual subtree when we remove it.
we have an identical stopping condition as in the adding-a-subtree case --
we stop when we hit a value of higher precedence (since that element
will be the one referenced further down the tree).