use a vtable for providers instead of an array.
this is mostly to clean up our usage of providers.  we used to index
into an array, cast, etc.  now we just use a field in the vtable.

also, we were overwriting FrameworkElementProvider for several of the
dynamic providers provided by subclasses.  make sure we delete the
previous dynamic provider.