remove checks if the autocreate provider is there, and only use the inherited provider where needed.
we had lots of if (providers.autocreate), when we unconditionally create
it in the ctor.  remove the checks.

also, not all DependencyObject subclasses need the inherited provider.
In fact only UIElement and TextElement subclasses need it.  This could
speed things up for all the DO's that don't subclass from those (i.e.
EventArgs, among other things.)