add initial support for implicit style.
we piggyback the default style stuff.  when creating a frameworkelement
(not just a control anymore) we search for the implicit style.  This
falls back to using the generic.xaml resources so we get the default
style for free.  whenever a FWE is loaded we search again for the
implicit style, so we pick up styles in resources in the visual
tree/application resources as well.

The code doesn't currently handle adding a new implicit style somewhere
in the hierarchy to override the current one, and it doesn't handle the
case where the implicit style is removed from the resources.