[Debug] Leak report now displays managed heap size and immortal objects
The leak report will now by default display the size of the managed
heap. It also has a new mode which modifies the output to only show
immortal objects, objects with a refcount >=2 which are gchandled.

There is also an extra debug printf to catch the case where an object
has a refcount of 1 and still has a strong gchandle.