[Clocks] Ensure that clock groups always trigger their children
If we have a Storyboard whose only timeline has a duration of 0, we never actually start or complete the children of that storyboard. The first call to UpdateFromParentTime doesn't update the children as the current state of the storyboard is Stopped. We then immediately complete the storyboard and enter the Filling state. As a result, all subsequent UpdateFromParentTime calls don't iterate the children, so they are never invoked. Fixes part of drt 2102.