[Selector] Update the selection logic to take MultiSelect into account
During the SL2 era it was trivial to select/unselect items using a single method call as we never had to worry about keeping multiple items selected at the same time. With SL3+ we need to handle multiple selected items at the same time so we now need to be explicit when deciding whether or not an element should be selected or unselected. Expose several additional methods on the Selection class to allow the caller to correctly select/unselect/replace the existing selection. Fixes parts of cs1.