[Selector] When refreshing the selector properties, use the correct item
Usually when Select is called on an item which is already selected, it's because the user tried to select an item which doesn't exist and so we're re-selecting the old item to refresh the properties. In this scenario, we need to ensure that we use the existing 'SelectedItem' to refresh the SelectedIndex and SelectedValue properties otherwise the Selector properties may reflect the values for the second or subsequent selected item, not the first selected item. Fixes the last of cs1.