[Image] Don't emit failed events if the image CRC is corrupt
DRT 0 expects that no error will be thrown if a png is loaded with an invalid
CRC. DRT 358 expects that if an image can't be loaded, the Failed event is
emitted. This patch attempts to clearly separate these two conditions so that
we emit the event only at the correct time. This fixes a regression in #0 while
still allowing #358 to succeed.