fix up a lot of the inherited provider stuff.  partial fixes for DRT cfd1.
allow DO's to override inheritance, and refuse to participate on a
property-by-property basis.  Image and MediaElement do this with
FlowDirection.

also, remove the geometric expansion of tree walks we run into when
setting an inherited property on the root of a tree.  This speeds up
cfd1 from the neighborhood of 3 hours to about 5 minutes.