more fixes for DRT cfd1 (and the logical tree in general)
when we clear the ItemsCollection the ItemsControl needs to know about
it so it can call SetLogicalParent(null) on each of its dependencyobject
children.

Fix this by removing the INotifyCollectionChanged interface from PFC<T>
and exposing the unmanaged collection changed interface there -- the
unmanaged collection changed interface allows us to notify on "Clearing"
(the state *before* the clear happens), while the INotify* interface
only allows .Reset (which is the state *after* the clear happens.)

Add back in the mapping between unmanaged/INotify* interfaces into the
two collections that need it for the automation peers to work
(ItemCollection and UIElementCollection).