[Media] Emit errors at the correct time.
DRt 171 and 173 both rely on the fact that setting an invalid URI on an Image
or MediaElement does not emit a synchronous Failed event. The event must be
emitted 1 tick later, otherwise when parsing we will invoke the javascript
OnError function rather than any event handler which is attached to the element.
i.e. <Image Source=uri ImageFailed='SomeHandler' /> fails to invoke SomeHandler
without this patch.