[Inheritance] pretty much completely rework the way we do inheritance.
We now make one pass over subtrees instead of multiple, and the
recursive methods are limitted to WalkTree/WalkSubTree.  All inherited
property propagation is done with these methods - DependencyObject::ProviderValueChanged
doesn't do any communication down the tree if the precedence is
Inherited.

This allows us to keep a simple InheritedContext class which can be used
to hold a stack of states about properties/sources as we traverse.  It
also gives us a solid way to know when we can *stop* traversing the
tree -- if all sources are different from those that we started with
(meaning we've picked up higher precedence sources along the way) we can
stop.