[Animations] Use a value coercion callback for KeyTime properties
turns out SetValue (*KeyFrame.KeyTime, null) is allowed (even though the
property is a value type and not nullable), but leads to an exception
when you Begin the storyboard.

we need to pass nulls to unmanaged, so remove the duplicate check in
managed.  We supply a KeyTime Coerce callback which returns an invalid
(but non-null) KeyTime value when you SetValue (..., null).  When
resolving keyframes, return false (invalid) if any of the keytimes are
invalid.

fixes drt #114