[ItemCollection] Revert ca1eb3 and fix it another way.
We currently can't propagate exceptions through events so by doing actual
work inside the native CollectionChanged event we can end up dropping
exceptions which should be propagated to the user. The simplest way to fix
this is to just add a Clearing event and raise it manually like we do for
the rest of the INCC events. This unregresses ComboBoxTest.ContainerItemTest4.