[Storyboard] Running storyboards should not be not eligible for GC
When a storyboard is running, it should not be eligible for GC. DRT
537 is breaking because they start storyboards without holding a
reference to them and so we GC them. Fix this by reffing in Begin and
unreffing in Stop. Note that we keep track of whether or not we've
reffed using a bool to simplify the case where Begin is called multiple
times in a row.