[HtmlBridge] Use ScriptObject serialization methods on System.ServiceModel.Web
On Moonlight, ScriptObject types have custom (de)serialization methods that
need to be called whenever a user object gets serialized by
System.Runtime.Serialization.Json and includes properties of that type.
This configures the new registration method for the tuner and calls it for
ScriptObject. Also, adds a -d:DESKTOP define so it only gets called
for the browser build, desktop doesn't support this (no browser)

Bumps mono revision
Fixes drt 983 for good