[VSP] Fix up the arguments passed to children in Measure/Arrange
We should not limit the size passed to the children based on the owning
containers MaxWidth/MaxHeight values. When VSP is scrolling we explicitly
want to pass infinity. In the non-scrolling case the available size will
already have been trimmed. We also need to ignore the desired size when
performing the Arrange as the element may be set to stretch to fill and
could have a desired size of 0,0.