Hold on to XamlContext parent refs.
Its possible that a context's parent can be freed but the context
hasn't been freed yet, this causes a crash when we try to lookup
items, because named item lookup does a recursive search up the tree.

We now ref the parent when we create a new xaml context and unref it
when the new context is freed.