[OSX] Implement the initial PAL backend for native rendering on OSX
This is a young implementation of the PAL backend for Moonlight.
It can render animations, text, probably transforms if built will gallium
support.  Mouse events are fully hooked up, the keyboard is about 80%
done.  Check README.darwin for a rough TODO list and instructions
on how to get things started.