[Loaded] Ensure we remove all loaded handlers from the Deployment
It's possible to queue up the same loaded delegate to be invoked multiple
times by adding and removing an element from the live tree multiple times
in a row. In this scenario if the loaded handler removes itself when it is
invoked and a GC happens before the subsequent invocations of the delegate,
we can end up invoking a GC'ed delegate. We can fix this by ensuring that
'RemoveAllMatchingHandlers' does not bail out after it finds one matching
handler. I also added an additional sanity to ensure that 'ForeachHandler'
ignores event closures which are pending removal.