[RichTextBox] correct a few problems with InlineUIContainer layout.
if the UIElement has a BaselineOffsetProperty (TextBlock, TextBox,
PasswordBox, RichTextBox), use that to line up the text with that on the
line.  Also add the font descend to the uielement height to calculate the
line height, so uielements line up with their bottom edge on the
baseline if they don't have a baseline offset.