[VSM] Correctly detect (UIElement.Width) and Width as being the same
If someone declares a VisualState which targets the "Width" property
and a VisualTransition which targets "(UIElement.Width)", we need to
resolve these property paths to allow VSM to check if the storyboards
are targetting the same DP on the same DO. This allows us to do
correct duplicate checking when attempting to figure out if we need to
create dynamic storyboards for transitions or not. Fixes drt 892 and
some new moon-unit tests.