[Leak] Don't clear the managed peer callbacks.
All of the managed peer callbacks are static and check that the
managed peer is alive before invoking the callback. As such, there's
no need to NULL them out when a DOs finalizer is run.

The main reason for this change is that it's possible for native land
to verify the callback is not null and be in the process of invoking
it when the finalizer thread decides to call ClearManagedCallbacks.