[Leak] Don't detach expressions in the finalizer thread
We can't detach expressions like this as it can cause DP PropertyChanged
events to be emitted on our finalizer thread and blow things up. There's
also a distinct possibility that if a subtree is being GC'ed, the parent
of this element may have already started its unref_delayed while the
managed finalizer is running on the element which can cause all sorts
of nasty.