[LeakFix] TemplateBindings need to use the WeakEventListener pattern
When a template element attaches a TemplateBinding to the template
owner, we need to store that delegate on the template owner just like we
do when attaching to regular events. Otherwise we can end up invoking a
GC'ed delegate if the template item is GC'ed and then the property is
changed on the template owner.

Normally this would then result in the template item being kept alive
as long as the template owner is alive. This is pretty undesireable
so instead of directly attaching the event, use a WeakEventListener.
This way we can attach to property change events and allow the
template item to be GC'ed.