Implement a new profiler to diagnose GCHandle related immortality
It's now possible to track what objects are being kept alive by strong
GCHandles and also see the stacktrace for every gchandle that was not
freed for a specific type.

The profiler can be enabled by exporting: MOON_PROFILER=gchandle .
On application shutdown, a GCHandle 'leak' report will be printed before
the appdomain is shut down. This will show all the objects being
retained because they have a strong gchandle attached to them.

If you want to see where the strong gchandle was allocated, all you have
to do is export: GCHANDLES_FOR_TYPE=TYPENAME. I.e.

export GCHANDLES_FOR_TYPE=Uri

If multiple strong handles are targetting the same CLR object, these
will be grouped together. This makes it easy to see if 10,000 handles
are targetting the same object or if its 10,000 different objects.