[Leak] Allow GCHandles to use add/clear strong ref
If we store objects in native with a strong GCHandle we make them,
and everything they reference in their object graph, immortal until
that GCHandle is freed. It's fairly trivial to end up in a scenario
where the gchandled object in the Value* references the DO the value*
is stored on, therefore making the entire subtree immortal. Fix this
by modifying the AddStrongRef/ClearStrongRef machinary to allow
both EventObjects and regualar gchandled clr objects.