[Leak] Prevent clocks and their targetted DOs from being immortal
DependencyObjects do not keep a strong reference to the storyboards
which are
animating them and the storyboards do not keep a strong reference to the
DOs they animate.

Storyboards are forcibly kept alive while they are running. We use
Deployment::KeepAlive to ensure that the storyboard is not gc'ed early
and remove the keepalive when the associated clock is no longer in the
Active state. This is similar enough to the behaviour of SL that we no
longer leak in the common cases like VisualStateManager.