Don't manually specify the Type::Kind when calling DOHelper.Lookup
There are a few places where we hard code a specific Type::Kind when
looking up the managed peer. This is bad as we're assuming that the
return value in native will always be of that type and in some cases
it's incorrect. This meant we generated an incorrect wrapper object.