Remove the MediaAttributeCollection managed peer
This is a native only collection so there's no point in having a
managed peer for it. By not creating a managed peer, we don't have to
worry about accidentally passing the object into addStrongRef which
means we won't accidentally hold a weak ref in native when there is
no corresponding managed ref. This can happen because this DO is
transformed into a Dictionary<string, string> in Value.ToObject.