[Leak] Don't keep DependencyObject.Parent in managed land
This value shouldn't be stored in managed as it can keep objects
alive longer than it should. For example if you set the foreground
of a control to a SolidColorBrush, the control will be kept alive
as long as the brush lives because the Brush.Parent property is the
control.