[Leak] Ensure the native XamlContextInternal refs its resource_list
We can't store raw pointers in a GSList without reffing them as otherwise
they can be GC'ed from beneath us and we'll just end up invoking functions
on freed memory. Eugh.

The fix is to ref the objects stored in the resource list, but this has a
very very high risk of creating immortal objects. Unfortunately other than
restructuring the native parser to be capable of holding weakrefs, there's
no other way of fixing this.

Fixes the crasher in #658929.