[Leak] Unregister event handlers attached by ImplicitStyleProvider
We need to unregister the template_detached handler when disposing
the ImplicitStyleManager as the templates will live longer than
the ImplicitStyleProvider and if the template is detached after the
provider has been GC'ed we'll invoke GC'ed memory.

I also added back in the listener for the destroyed event on the
template but I'm not sure if this is required or not. I figured it'd
be safest for now to add it in. Fixes the crasher reported in #658929