[Leak] Attempt to fix a rare crash in drt 291.
It's possible that the timeline associated with some clocks has
already been freed before the corresponding AnimationStorages
are torn down. As a result, if there are multiple storages on the same
DP, we blow up. When we detach the first storage we enable the second
one, which then tries to get the current value off the timeline. However
if this has been GC'ed we crash. Attempt to fix this by storing the timeline
in a WeakRef.