Media: Fix support for StartTime when starting media.
The demuxer now handles any StartTime we may have, by seeking to it when
opening the media. The demuxer also handles any Duration we may have, by
just treating it as a end-of-stream for all streams. The pipeline keeps
track of the real position, and MediaElement subtracts the StartTime in
all user-visible properties (Markers, CurrentPosition, Duration, etc).