Pipeline: create two frame queues and improve seeking.
Create two frame queues: one queue with demuxed frames and one queue with
decoded frames. We're reporting buffering progress based on the queue of
demuxed frames, while we keep the queue of decoded frames smaller to use
(a lot) less memory. Also make seeking code a bit better by adding a seek
generation field to frames, and if we get a frame from the wrong generation
we know immediately that we can drop it.