[Perf] Cache and optimise getting/setting properties in the parser
Property lookups are now done with a Dictionary cache to speed them up.
We also attempt to generate delegates using expression trees to give us
a getter/setter delegate which can be invoked with ~0 overhead. If that
can't be done, we fall back to invoking from the PropertyInfo as before.

Overall this gives us another 30%+ speedup when parsing templates.