[Perf] Remove the NativeDependencyObjectHelper.objects dictionary
We can get the managed object for any native EO pointer by calling
into native and returning the target of its GCHandle. This means we
don't have to store a second GCHandle in managed for each EO and we
don't have to have a massive dictionary mapping from native object to
managed object. Faster and less memory, hurrah!