[Leak] Prevent elements with ToolTips being retained forever
A static Dictionary<UIElement, ToolTip> will cause both the tooltip and
the UIElement the tooltip is on to be retained forever and ever and
ever. We only want this so we can map from UIElement -> Tooltip so
instead just store the ToolTip object on the UIElement using a custom
DependencyProperty. Fixes the leak in #656621.