[ItemsControl] ItemCollection should be a regular DP so it can be parented
We need to store ItemCollection in a DP so it can be parented by its owning
ItemsControl and then we need to ensure it parents DependencyObjects which are
added to it. This means that DOs which are added to an ItemCollection will need
to exist inside two collections, the ItemCollection itself and then the Panel
that the ItemsControl uses to display items. To avoid getting exceptions when
an item exists in both collections, we need to modify DependencyObjectCollection
to allow it to not-parent items if they are already parented by the ItemCollection.
This fixes part of drt 2101