[Leak] Do not use strings for the [Add|Clear]ManagedRef callback
We can't use strings for our managed ref holding machinery. Multiple
distinct DPs can be registered with the same name which leads to
collisions. It's also possible for a DP to be registered with the same
name as one of our native WeakRefs which again would lead to collisions
and cause us to GC elements earlier than they should be.

We can fix this by passing an opaque pointer from native instead of a char*.
This pointer is either the DependencyProperty itself (so no collisions are
possible), or in the case of a native WeakRef, it's an autogenerated opaque
 ID. This will only give us collisions if we allocate a DP in the memory
range of 1-30. Highly unlikely ;)

The result of this is that now it's impossible to clobber our memory
management with a cunningly devised DP registration.