[Leak] Don't crash when enabling a previous AnimationStorage.
When a storyboard is being torn down it disables its animation storage. If
there are two or more activate timelines on the same DP this can result in
us attempting to activate a previous animation storage. This can lead to a
crash if the timeline associated with the previous animation storage has been
GC'ed and has been partially torn down.

To prevent a crash, we need to cache the current_value in the animation storage
so we don't have to access the associated timeline to get it. When calling Enable
on an animation storage, we just use the current_value. When the clock ticks, we
update that value and apply it.

Prevents leak test 28 from crashing.