[Leak] Proxy loaded handlers in a safer manner.
The LabelTests drt crashes 100% of the time due to invoking collected
Loaded event handlers. I can't for the life of me figure out what
pattern of adding/removing/proxying loaded handlers causes the issue, so
work around it by never storing the Loaded delegates on deployment.

Instead of storing the loaded delegate on the Deployment we store the
token from when the handler was added to its UIElement. When deployment
invokes its Loaded event, we then try to emit the eventhandler with that
token on the original UIElement.