[Binding] Fix two way bindings to DPs when a type converter is needed
When updating a two way binding to an attached DP we need to use
PropertyPathNode.ValueType instead of assuming there's a PropertyInfo
and reading the PropertyType from that. In the case of a DP binding
with no clr wrapper, there will be no PropertyInfo. ValueType is stored
precisely for this reason. Fixes the StringIndexerPlatformManagedAttachedDPTest
subtest in drt 233.