[Focus] Handle GotFocus/LostFocus events correctly when focusing in/out
If the plugin gets a FocusIn event and no element is focused, it should attempt to focus an element. If there's already a focused element, it should just re-emit the GotFocus event on that element. We also have to not emit GotFocus/LostFocus events when toggling to/from fullscreen. Fixes parts of drt tn1.