[TabNavigation] Fix the remainder of DRT tn1
When tabbing to the next item, we need to find the parent Control of the currently focused Control and walk its entire subtree to find all potential tab targets. This is the list we use to figure out which control to tab to next. We also need to treat elements which have no TabIndex set as if they had a the default TabIndex value of 0. Fixes the remainder of drt tn1.