[ContentPresenter] Null the DataContext after clearing the root
If we have a twoway databinding against TextBox.Text, the binding will
update when the TextBox loses focus. If we clear the DataContext before
the focus event is emitted, the binding will be broken (no data source)
by the time the LostFocus event is raised and it won't update right.

This fixes testcase C and D in #667657.