[Selector] Try to keep the SelectedItems in Selector and Selection in sync
If we manually add an item to ItemsControl.SelectedItems that is not allowed
in the ItemsControl.Items collection, we need to remove it from SelectedItems.
The logical place to do this is within the Change handler, but this throws an
exception. Attempt to work around this by keeping a boolean which tracks if the
collection is stale and if so, update it later when it's safe to do so. This
matches the MS behaviour from what I can tell. Fixes the failing tests in
moon-unit and doesn't appear to regress anything in 233.