[Bindings] Fix an issue where we potentially attach 1000's of Loaded handlers
Rewrite the binding code slightly to fix a number of issues. Firstly remove
the 'DataSource' property and convert it to a function which is called once
when the binding is attached. Previously every time we called this we could
end up attaching more and more loaded handlers.

Secondly, OneTime bindings do not listen for property changes on the bound
item. However, if they're bound to the DataContext, they listen for when the
DataContext changes and update then. With this taken into account inside the
binding code, we can now (finally!) remove the code in FrameworkElement to
forcibly invalidate all bindings every time the datacontext changes.

This apparently causes no regressions and fixes the issue in
https://finanse.ifin24.pl where we added thousands of eventhandlers.