[Binding] Fix up usage of DependencyProperty.DefaultValue
We need to use the DP default value if there's no fallback value provided
and the binding is broken. We also need to use it if we databind to the
datacontext and the datacontext is null. This case does not count as a
broken binding so we don't hit the fallbackvalue/dp default value path.