[Binding] Fix the interaction of mentors with DataContext based bindings
Bindings on FrameworkElements must always listen to either the local DataContext
or the DataContext of it's visual parent. We explicitly use the VisualParent as
opposed to FrameworkElement.Parent as DataContext changes are propagated along
the visual tree in native and we also have no guarantee that a FrameworkElement
Parent will ever be set (e.g. when expanding ContentControl.ContentTemplate).

The DataContext on the Mentor is only used if the binding is attached to an
element which is *not* a FrameworkElement, e.g. SolidColorBrush. This fixes