[CollectionView] Don't defer refresh when we actually refresh
Several properties and methods in the CollectionView implementation must
throw InvalidOperationExceptions if accessed while refresh is deferred.
As such we can no longer use the trick of increasing the defer level
to avoid emitting events when we actually being a Refresh. Instead use
another boolean to indicate when we should ignore changes in the filtered
list.