[PropertySystem] Split up default values from autocreated values
A single instance of a Value* for a default value can be used by every
DependencyObject. For example there is no need for every single
FrameworkElement to create a Value (Double.NaN) for it's width and height
properties, it can just use the actual Value* stored on the DP itself.
By splitting AutoCreatePropertyValueProvider into a separate
DefaultValueProvider and AutoCreateProvider again we can get this benefit.