[ref] Change value_free_value calls to Value.Dispose calls
If we get a native Value* pointer from native we are supposed to use
value_delete_value to free it. If we have a managed Value we should just
call Value.Dispose () which in turn calls value_free_value. Calling
value_free_value on a native pointer is more than likely a bug so to
help catch these cases refactor the code so that the only usage of
value_free_value is in Value.Dispose ().