Substantially rework the way implicit styles are handled.
move the tree traversal code to managed for when a style changes in
either application resources or visual tree resources.  also change
things such that we don't reapply/revalidate all styles when only 1 has
changed.

The tree walking code in managed also now applies the style directly
(and only to those objects which it should apply to), instead of forcing
all frameworkelements to walk the tree back up to the root looking for
an applicable style.  this should speed things up a bit.