[Parent] Handle namescoping more correctly for elements with multiple parents
Elements which support multiple parents should only have their name registered
in the first (or primary) parent. Any subsequent (or secondary) parents which
are added to the element do not register the elements name in their namescope.
The one exception to this (so far) is elements which are in an ItemsCollection
and also in a Panel. In this scenario we need to register in both namescopes.

Special care has to be taken when registering/unregistering names as names can
only be registered/unregistered when there are no secondary parents. See tests
in LogicalTreeTest.cs for all the bizarre corner cases I've thought up.