[Binding] More touchups to how we handle ElementName bindings
The rules for deciding what element to look up appear to be something
like: Choose Binding.Target if it's a FrameworkElement otherwise choose
Binding.Target.Mentor. If the name cannot be found on the chosen element
and the TemplateOwner of that element is non-null, do a final lookup
using that.

This fixes the new moon-unit test, the presenter video on the
http://www.cloudops.de/event/videos/ site and doesn't regress telerik
or DRT 232.