BitmapImage: ensure we emit the correct error when an invalid image is received
BitmapImage: when receiving an invalid image ensure that we emit the 'invalid
image' error, and not the 'download failed' error due to aborting the
downloader (which we do since the image is invalid). So create moon_error
before aborting, and don't emit any 'download failed' errors if we already
have a moon_error. Fixes regression in #641 (the exact issue is that we ended
up emitting both events, the first causing all handlers to be removed, then
the second reached the plugin error handler and failed the test).